upload/koji: move metadata structures to a separate file
Increase readability by logically splitting the build metadata
structures from the actual upload API implementation.
